+++ test/viking/view/form_builder.js
@@ -22,79 +22,61 @@
         equal(form.options, options);
     });
 
-    // render()
-    // ========
-    test("#render()", function () {
-        var form = new FormBuilder(this.model, function() { return ''; })
-        
-        equal( form.render(), '<form method="get"></form>' );
-    });
-    
-    test("#render() - options: {method: get}", function() {
-        var form = new FormBuilder(this.model, {method: "get"}, function() { return ''; })
-        
-        equal( form.render(), '<form method="get"></form>' );
-    });
-    
-    test("#render() - options: {method: post}", function() {
-        var form = new FormBuilder(this.model, {method: "post"}, function() { return ''; })
-        
-        equal( form.render(), '<form method="post"></form>' );
-    });
-    
-    test("#render() - options: {method: patch}", function() {
-        var form = new FormBuilder(this.model, {method: "patch"}, function() { return ''; })
-        
-        equal( form.render(), '<form method="post"><div style="margin:0;padding:0;display:inline"><input name="_method" type="hidden" value="patch"></div></form>' );
-    });
-    
-    test("#render() - options: {method: put}", function() {
-        var form = new FormBuilder(this.model, {method: "put"}, function() { return ''; })
-        
-        equal( form.render(), '<form method="post"><div style="margin:0;padding:0;display:inline"><input name="_method" type="hidden" value="put"></div></form>' );
-    });
-    
-    test("#render() - options: {method: delete}", function() {
-        var form = new FormBuilder(this.model, {method: "delete"}, function() { return ''; })
-        
-        equal( form.render(), '<form method="post"><div style="margin:0;padding:0;display:inline"><input name="_method" type="hidden" value="delete"></div></form>' );
-    });
-    
-    test("#render() - options: {multipart: true}", function() {
-        var form = new FormBuilder(this.model, {multipart: true}, function() { return ''; })
-        
-        equal( form.render(), '<form enctype="multipart/form-data" method="post"></form>' );
-    });
-    
-    test("#render() - options: {multipart: true, method: patch}", function() {
-        var form = new FormBuilder(this.model, {multipart: true, method: 'patch'}, function() { return ''; })
-        
-        equal( form.render(), '<form enctype="multipart/form-data" method="post"><div style="margin:0;padding:0;display:inline"><input name="_method" type="hidden" value="patch"></div></form>' );
-    });
-    
-    test("#render() - options: {multipart: true, method: get}", function() {
-        var form = new FormBuilder(this.model, {multipart: true, method: 'get'}, function() { return ''; })
-        
-        equal( form.render(), '<form enctype="multipart/form-data" method="post"><div style="margin:0;padding:0;display:inline"><input name="_method" type="hidden" value="get"></div></form>' );
-    });
-    
     // checkBox()
     // ==========
     test("#checkBox() passes to Viking.View.Helpers.checkBox", function () {
         expect(5);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.checkBox
         
         Viking.View.Helpers.checkBox = function(m, attribute, options, checkedValue, uncheckedValue) {
             strictEqual(model, m);
             strictEqual('key', attribute);
-            strictEqual(1, options);
+            deepEqual({}, options);
             strictEqual(2, checkedValue);
             strictEqual(3, uncheckedValue);
         }
-        form.checkBox('key', 1, 2, 3);
+        form.checkBox('key', {}, 2, 3);
+        
+        Viking.View.Helpers.checkBox = old_func;
+    });
+    
+    test("#checkBox() uses namepsace on for attribute", function () {
+        expect(5);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.checkBox
+        
+        Viking.View.Helpers.checkBox = function(m, attribute, options, checkedValue, uncheckedValue) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'ns[model][key]'}, options);
+            strictEqual(2, checkedValue);
+            strictEqual(3, uncheckedValue);
+        }
+        form.checkBox('key', {}, 2, 3);
+        
+        Viking.View.Helpers.checkBox = old_func;
+    });
+    
+    test("#checkBox() allows name attribute to overridden", function () {
+        expect(5);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.checkBox
+        
+        Viking.View.Helpers.checkBox = function(m, attribute, options, checkedValue, uncheckedValue) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'overridden'}, options);
+            strictEqual(2, checkedValue);
+            strictEqual(3, uncheckedValue);
+        }
+        form.checkBox('key', {name: 'overridden'}, 2, 3);
         
         Viking.View.Helpers.checkBox = old_func;
     });
@@ -105,7 +87,7 @@
         expect(6);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.collectionSelect
         
         Viking.View.Helpers.collectionSelect = function(m, attribute, collection, valueAttribute, textAttribute, options) {
@@ -114,9 +96,49 @@
             strictEqual(1, collection);
             strictEqual(2, valueAttribute);
             strictEqual(3, textAttribute);
-            strictEqual(4, options);
+            deepEqual({}, options);
         }
-        form.collectionSelect('key', 1, 2, 3, 4);
+        form.collectionSelect('key', 1, 2, 3, {});
+        
+        Viking.View.Helpers.collectionSelect = old_func;
+    });
+    
+    test("#collectionSelect() obeys namespace", function () {
+        expect(6);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.collectionSelect
+        
+        Viking.View.Helpers.collectionSelect = function(m, attribute, collection, valueAttribute, textAttribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, collection);
+            strictEqual(2, valueAttribute);
+            strictEqual(3, textAttribute);
+            deepEqual({name: 'ns[model][key]'}, options);
+        }
+        form.collectionSelect('key', 1, 2, 3, {});
+        
+        Viking.View.Helpers.collectionSelect = old_func;
+    });
+    
+    test("#collectionSelect() allows name to be overridden", function () {
+        expect(6);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.collectionSelect
+        
+        Viking.View.Helpers.collectionSelect = function(m, attribute, collection, valueAttribute, textAttribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, collection);
+            strictEqual(2, valueAttribute);
+            strictEqual(3, textAttribute);
+            deepEqual({name: 'overridden'}, options);
+        }
+        form.collectionSelect('key', 1, 2, 3, {name: 'overridden'});
         
         Viking.View.Helpers.collectionSelect = old_func;
     });
@@ -127,15 +149,49 @@
         expect(3);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.hiddenField
         
         Viking.View.Helpers.hiddenField = function(m, attribute, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
-            strictEqual(1, options);
+            deepEqual({}, options);
         }
-        form.hiddenField('key', 1);
+        form.hiddenField('key', {});
+        
+        Viking.View.Helpers.hiddenField = old_func;
+    });
+    
+    test("#hiddenField() obeys namespace", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.hiddenField
+        
+        Viking.View.Helpers.hiddenField = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({name: 'ns[model][key]'}, options);
+        }
+        form.hiddenField('key', {});
+        
+        Viking.View.Helpers.hiddenField = old_func;
+    });
+    
+    test("#hiddenField() allows name to be overridden", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.hiddenField
+        
+        Viking.View.Helpers.hiddenField = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({name: 'overridden'}, options);
+        }
+        form.hiddenField('key', {name: 'overridden'});
         
         Viking.View.Helpers.hiddenField = old_func;
     });
@@ -146,16 +202,52 @@
         expect(4);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.label
         
         Viking.View.Helpers.label = function(m, attribute, content, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
             strictEqual(1, content);
-            strictEqual(2, options);
+            deepEqual({}, options);
         }
-        form.label('key', 1, 2);
+        form.label('key', 1, {});
+        
+        Viking.View.Helpers.label = old_func;
+    });
+    
+    test("#label() uses namepsace on for attribute", function () {
+        expect(4);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.label
+        
+        Viking.View.Helpers.label = function(m, attribute, content, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, content);
+            deepEqual({'for': 'ns_model_key'}, options);
+        }
+        form.label('key', 1, {});
+        
+        Viking.View.Helpers.label = old_func;
+    });
+    
+    test("#label() allows for attribute to be overridden", function () {
+        expect(4);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model)
+        old_func = Viking.View.Helpers.label
+        
+        Viking.View.Helpers.label = function(m, attribute, content, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, content);
+            deepEqual({'for': 'me'}, options);
+        }
+        form.label('key', 1, {'for': 'me'});
         
         Viking.View.Helpers.label = old_func;
     });
@@ -166,15 +258,49 @@
         expect(3);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.passwordField
         
         Viking.View.Helpers.passwordField = function(m, attribute, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
-            strictEqual(1, options);
+            deepEqual({}, options);
         }
-        form.passwordField('key', 1);
+        form.passwordField('key', {});
+        
+        Viking.View.Helpers.passwordField = old_func;
+    });
+    
+    test("#passwordField() obeys namespace", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.passwordField
+        
+        Viking.View.Helpers.passwordField = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'ns[model][key]'}, options);
+        }
+        form.passwordField('key', {});
+        
+        Viking.View.Helpers.passwordField = old_func;
+    });
+    
+    test("#passwordField() allows name to be overridden", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.passwordField
+        
+        Viking.View.Helpers.passwordField = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'overridden'}, options);
+        }
+        form.passwordField('key', {name: 'overridden'});
         
         Viking.View.Helpers.passwordField = old_func;
     });
@@ -185,16 +311,52 @@
         expect(4);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.radioButton
         
         Viking.View.Helpers.radioButton = function(m, attribute, tagValue, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
             strictEqual(1, tagValue);
-            strictEqual(2, options);
+            deepEqual({}, options);
         }
-        form.radioButton('key', 1, 2);
+        form.radioButton('key', 1, {});
+        
+        Viking.View.Helpers.radioButton = old_func;
+    });
+    
+    test("#radioButton() obeys namespace", function () {
+        expect(4);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.radioButton
+        
+        Viking.View.Helpers.radioButton = function(m, attribute, tagValue, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, tagValue);
+            deepEqual({'name': 'ns[model][key]'}, options);
+        }
+        form.radioButton('key', 1, {});
+        
+        Viking.View.Helpers.radioButton = old_func;
+    });
+    
+    test("#radioButton() allows name to be overridden", function () {
+        expect(4);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.radioButton
+        
+        Viking.View.Helpers.radioButton = function(m, attribute, tagValue, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, tagValue);
+            deepEqual({'name': 'overridden'}, options);
+        }
+        form.radioButton('key', 1, {name: 'overridden'});
         
         Viking.View.Helpers.radioButton = old_func;
     });
@@ -205,16 +367,52 @@
         expect(4);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.select
         
         Viking.View.Helpers.select = function(m, attribute, collection, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
             strictEqual(1, collection);
-            strictEqual(2, options);
+            deepEqual({}, options);
         }
-        form.select('key', 1, 2);
+        form.select('key', 1, {});
+        
+        Viking.View.Helpers.select = old_func;
+    });
+    
+    test("#select() obeys namespace", function () {
+        expect(4);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.select
+        
+        Viking.View.Helpers.select = function(m, attribute, collection, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, collection);
+            deepEqual({'name': 'ns[model][key]'}, options);
+        }
+        form.select('key', 1, {});
+        
+        Viking.View.Helpers.select = old_func;
+    });
+    
+    test("#select() allows name to be overridden", function () {
+        expect(4);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.select
+        
+        Viking.View.Helpers.select = function(m, attribute, collection, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            strictEqual(1, collection);
+            deepEqual({'name': 'overridden'}, options);
+        }
+        form.select('key', 1, {name: 'overridden'});
         
         Viking.View.Helpers.select = old_func;
     });
@@ -225,15 +423,49 @@
         expect(3);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.textArea
         
         Viking.View.Helpers.textArea = function(m, attribute, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
-            strictEqual(1, options);
+            deepEqual({}, options);
         }
-        form.textArea('key', 1);
+        form.textArea('key', {});
+        
+        Viking.View.Helpers.textArea = old_func;
+    });
+    
+    test("#textArea() obeys namespace", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.textArea
+        
+        Viking.View.Helpers.textArea = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'ns[model][key]'}, options);
+        }
+        form.textArea('key', {});
+        
+        Viking.View.Helpers.textArea = old_func;
+    });
+    
+    test("#textArea() allows name to be overridden", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.textArea
+        
+        Viking.View.Helpers.textArea = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'overridden'}, options);
+        }
+        form.textArea('key', {name: 'overridden'});
         
         Viking.View.Helpers.textArea = old_func;
     });
@@ -244,17 +476,69 @@
         expect(3);
         
         var model = this.model;
-        var form = new FormBuilder(this.model, function() { return ''; })
+        var form = new FormBuilder(this.model)
         old_func = Viking.View.Helpers.textField
         
         Viking.View.Helpers.textField = function(m, attribute, options) {
             strictEqual(model, m);
             strictEqual('key', attribute);
-            strictEqual(1, options);
+            deepEqual({}, options);
         }
-        form.textField('key', 1);
+        form.textField('key', {});
         
         Viking.View.Helpers.textField = old_func;
+    });
+    
+    test("#textField() obeys namespace", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.textField
+        
+        Viking.View.Helpers.textField = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({name: 'ns[model][key]'}, options);
+        }
+        form.textField('key', {});
+        
+        Viking.View.Helpers.textField = old_func;
+    });
+    
+    test("#textField() allows name to be overridden", function () {
+        expect(3);
+        
+        var model = this.model;
+        var form = new FormBuilder(this.model, {namespace: 'ns'})
+        old_func = Viking.View.Helpers.textField
+        
+        Viking.View.Helpers.textField = function(m, attribute, options) {
+            strictEqual(model, m);
+            strictEqual('key', attribute);
+            deepEqual({'name': 'overridden'}, options);
+        }
+        form.textField('key', {name: 'overridden'});
+        
+        Viking.View.Helpers.textField = old_func;
+    });
+    
+    // formFor()
+    // =========
+    test("#fieldsFor() yields a FormBuilder with the namespace set", function (attribute, options, content) {
+        var formBuilder = new FormBuilder(this.model);
+        
+        formBuilder.fieldsFor('key', function(f) {
+            equal(f.options.namespace, 'model');
+        });
+    });
+    
+    test("#fieldsFor() works with a FormBuilder that already has a namespace", function (attribute, options, content) {
+        var formBuilder = new FormBuilder(this.model, {namespace: 'ns'});
+        
+        formBuilder.fieldsFor('key', function(f) {
+            equal(f.options.namespace, 'ns[model]');
+        });
     });
     
 }());